Resolve #2 - Make the repository easier to set up and install
Also migrates from tslint to eslint, upgrades some type dependencies, and fixes some
new errors produced by eslint. Prettier config is now added.

lintstaged and husky are added to avoid having to lint and format the entire code base
on each commit. This is a subjective change which could be reversed later.

eslint caused a lot of errors and warnings. Not all errors could be addressed in this
set of changes, so the new rules exclude several useful rules. This should be addressed
later on in another issue.