Use :before for border, enlarge font to stop input-zooming on mobile
These changes fix issues on mobile devices.

Mobile browsers have terrible support for box-shadow, so it seems like
trouble to use it. Instead there's a pseudo :before element which
accomplishes the same thing. Although it works better so far, it needs
more testing.

The font-size was set to 14px on the input, which causes mobile safari
to automatically zoom in to enlarge the text. It's an a11y thing which
can't be turned off, so the font needs to be larger. For some reason,
turning off user scaling doesn't help. This could be a simulator bug.

The min-height of the input element was moved to its container, and now
all items in the container have display: flex. This makes it so all
items fully fill the container in more browsers, not just the input.