Add initial mobile search selector
This provides some shared structure around search selector components and their data, making it so current selections can be reflected without using state. This works in both in mobile and desktop versions.

A significant todo was introduced to do with localizing `<option>` text. We currently use hooks to access state selectors, but that isn't possible here because the value of the selector needs to be accessed by a react component. The child of an `<option>` can only be a text node, so the output ends up as `[object Object]` - the `toString` value of the `<Text>` component in our case. I dug around a bit to find nicer ways to handle this, but for the moment I'm going to leave it as it is.

This also corrects a minor issue in the footer in which a <Link>'s to attribute wasn't internal.